logthrdestdrv: attempt to keep the batch size consistent accross retries
When we retry a batch, we just replay the messages to the insert() method.
However, if the meanwhile new messages have arrived in the queue,
we would add additional messages to the same, previously failing, batch.

That could mean that some of the messages are not actually retried enough
times, in extreme cases it could mean that messages are never
even retried.

To resolve this situation, try to keep the batch size consistent, by
not allowing more items in the batch than previously.

There could still be variations, when insert() decides to flush earlier,
but in those cases we would retry some items more times than configured
and not less.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>